requirements.txt
requests==2.33.1
google-cloud-storage==3.10.1
google-cloud-bigquery==3.41.0
streamlit==1.43.2
pydeck==0.9.1
pandas==2.2.3
plotly
pytz
db-dtypes
pyarrow
confluent-kafka==2.6.1
